backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
python-dotenv==1.0.0
google-generativeai==0.8.5
